pyproject.toml
[project]
name = "sre-copilot"
version = "0.1.0"
description = "AI-powered SRE Copilot Agent for incident triage and diagnosis"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.0",
    "langchain-openai>=0.3.0",
    "langchain-community>=0.3.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0.0",
    "docker>=7.0.0",
]

[project.optional-dependencies]
slack = [
    "slack-bolt>=1.18.0",
    "slack-sdk>=3.27.0",
]
dev = [
    "pytest>=8.0",
    "ruff>=0.8.0",
]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"